+++ README.md
@@ -83,9 +83,8 @@ client = pornhub.PornHub(keywords)
 
 for video in client.getVideos(10, page=2, full_data=True):
   print(video)
-  print(video['upload_date'])
+  print(video["upload_date"])
 ```
-
 
 #### Get more information about the single video
